requirements.txt
aiogram
pydantic
aiohttp
asyncio
requests
pandas
scrapy
json
bs4
aiomysql
mysql-connector-python
pydantic_settings
